+++ test/multiapi/Expected/AcceptanceTests/MultiapiNoAsync/multiapinoasync/_multiapi_service_client.py
@@ -65,12 +65,10 @@ class MultiapiServiceClient(MultiapiServiceClientOperationsMixin, MultiApiClient
         **kwargs  # type: Any
     ):
         if not base_url:
-            base_url = 'https://management.azure.com'
+            base_url = 'http://localhost:3000'
         self._config = MultiapiServiceClientConfiguration(credential, **kwargs)
         self._client = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
         super(MultiapiServiceClient, self).__init__(
-            credential,
-            self._config,
             api_version=api_version,
             profile=profile
         )